Move content from etc to by_score
